src/requirements.txt
discord.py
aiohttp
aiosqlite
requests
tzdata
fastapi
pymongo
motor
uvicorn[standard]
python-dotenv